# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_MeshCore-Beacon_beacon-server_internal_api.CrossIATAHop:
    properties:
      fromIata:
        type: string
      fromNode:
        allOf:
        - $ref: '#/definitions/github_com_MeshCore-Beacon_beacon-server_internal_api.ResolvedNode'
        description: last node in source IATA
      lastSeen:
        description: epoch ms
        type: integer
      toIata:
        type: string
      toNode:
        allOf:
        - $ref: '#/definitions/github_com_MeshCore-Beacon_beacon-server_internal_api.ResolvedNode'
        description: first node in target IATA
    type: object
  github_com_MeshCore-Beacon_beacon-server_internal_api.CrossIATARoute:
    properties:
      crossHop:
        allOf:
        - $ref: '#/definitions/github_com_MeshCore-Beacon_beacon-server_internal_api.CrossIATAHop'
        description: the boundary hop
      sourceSegment:
        description: route segment in source IATA
        items:
          $ref: '#/definitions/github_com_MeshCore-Beacon_beacon-server_internal_api.RouteHop'
        type: array
      targetSegment:
        description: route segment in target IATA
        items:
          $ref: '#/definitions/github_com_MeshCore-Beacon_beacon-server_internal_api.RouteHop'
        type: array
      totalHops:
        type: integer
    type: object
...
  github_com_MeshCore-Beacon_beacon-server_internal_api.KnownRoute:
    properties:
      firstSeen:
        description: epoch ms
        type: integer
      hopCount:
        type: integer
      hops:
        items:
          $ref: '#/definitions/github_com_MeshCore-Beacon_beacon-server_internal_api.RouteHop'
        type: array
      iata:
        type: string
      id:
        type: integer
      lastSeen:
        description: epoch ms
        type: integer
      observationCount:
        type: integer
    type: object
...
  github_com_MeshCore-Beacon_beacon-server_internal_api.PacketHeader:
    properties:
      payloadType:
        description: bits 2-5
        type: integer
      payloadTypeName:
        description: advert, request, group_text, etc.
        type: string
      payloadVersion:
        description: bits 6-7
        type: integer
      raw:
        description: hex-encoded single byte
        type: string
      routeType:
        description: bits 0-1
        type: integer
      routeTypeName:
        description: FLOOD, DIRECT, TRANSPORT_FLOOD, TRANSPORT_DIRECT
        type: string
    type: object
  github_com_MeshCore-Beacon_beacon-server_internal_api.PacketPathLength:
    properties:
      hashSize:
        description: per-hop hash size in bytes (1, 2, or 3)
        type: integer
      hopCount:
        description: number of path hashes present
        type: integer
      raw:
        description: hex-encoded single byte
        type: string
    type: object
  github_com_MeshCore-Beacon_beacon-server_internal_api.PacketTransportCodes:
    properties:
      regionCode:
        type: integer
      subRegionCode:
        type: integer
    type: object
...
  ? github_com_MeshCore-Beacon_beacon-server_internal_api.Page-github_com_MeshCore-Beacon_beacon-server_internal_api_PublicPacketSummary
  : properties:
      hasMore:
        type: boolean
      items:
        items:
          $ref: '#/definitions/github_com_MeshCore-Beacon_beacon-server_internal_api.PublicPacketSummary'
        type: array
      nextCursor:
        type: integer
    type: object
...
  github_com_MeshCore-Beacon_beacon-server_internal_api.PublicPacket:
    properties:
      channelHash:
        type: string
      decrypted:
        type: boolean
      firstHeardAt:
        type: integer
      firstToLastMs:
        type: integer
      header:
        $ref: '#/definitions/github_com_MeshCore-Beacon_beacon-server_internal_api.PacketHeader'
      lastHeardAt:
        type: integer
      observationCount:
        type: integer
      originPubkey:
        type: string
      packetHash:
        type: string
      resolvedRoute:
        items:
          $ref: '#/definitions/github_com_MeshCore-Beacon_beacon-server_internal_api.ResolvedHop'
        type: array
      scope:
        type: string
      transportCodes:
        $ref: '#/definitions/github_com_MeshCore-Beacon_beacon-server_internal_api.PacketTransportCodes'
    type: object
  github_com_MeshCore-Beacon_beacon-server_internal_api.PublicPacketLatestObserver:
    properties:
      displayName:
        type: string
      iata:
        type: string
      id:
        type: string
      pathBytes:
        type: string
      pathLength:
        $ref: '#/definitions/github_com_MeshCore-Beacon_beacon-server_internal_api.PacketPathLength'
      resolvedDestination:
        $ref: '#/definitions/github_com_MeshCore-Beacon_beacon-server_internal_api.ResolvedHop'
      resolvedPath:
        items:
          $ref: '#/definitions/github_com_MeshCore-Beacon_beacon-server_internal_api.ResolvedHop'
        type: array
      resolvedSource:
        $ref: '#/definitions/github_com_MeshCore-Beacon_beacon-server_internal_api.ResolvedHop'
    type: object
...
  github_com_MeshCore-Beacon_beacon-server_internal_api.PublicPacketSummary:
    properties:
      firstHeardAt:
        type: integer
      lastHeardAt:
        type: integer
      latestObserver:
        $ref: '#/definitions/github_com_MeshCore-Beacon_beacon-server_internal_api.PublicPacketLatestObserver'
      observationCount:
        type: integer
      packetHash:
        type: string
      payloadType:
        type: integer
      payloadTypeName:
        type: string
      routeType:
        type: integer
      routeTypeName:
        type: string
      scope:
        type: string
      summary:
        type: string
    type: object
...
  github_com_MeshCore-Beacon_beacon-server_internal_api.ResolvedHop:
    properties:
      confidence:
        description: '"high", "ambiguous", or "none"'
        type: string
      nodes:
        description: empty for "none", one for "high", multiple for "ambiguous"
        items:
          $ref: '#/definitions/github_com_MeshCore-Beacon_beacon-server_internal_api.ResolvedNode'
        type: array
      snr:
        type: number
    type: object
  github_com_MeshCore-Beacon_beacon-server_internal_api.ResolvedNode:
    properties:
      id:
        type: string
      latitude:
        type: number
      longitude:
        type: number
      name:
        type: string
      publicKey:
        description: hex-encoded prefix used for resolution
        type: string
    type: object
  github_com_MeshCore-Beacon_beacon-server_internal_api.RouteHop:
    properties:
      hashBytes:
        description: hex-encoded hash prefix
        type: string
      node:
        allOf:
        - $ref: '#/definitions/github_com_MeshCore-Beacon_beacon-server_internal_api.ResolvedNode'
        description: populated when node details are available
      nodeId:
        type: string
    type: object
...
  internal_api_handlers.APIError:
    properties:
      code:
        description: e.g. "not_found", "bad_request"
        type: string
      message:
        description: e.g. "channel not found"
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    name: MeshCore Beacon
    url: https://github.com/MeshCore-Beacon/beacon-server
  description: MeshCore network observation backend. Ingests LoRa packets from MQTT
    brokers, stores in PostgreSQL, and streams live events via WebSocket.
  license:
    name: AGPL-3-or-later
  termsOfService: https://github.com/MeshCore-Beacon/beacon-server
  title: MeshCore Beacon API
  version: 1.6.0
paths:
...
  /packets:
    get:
      parameters:
      - description: Filter by payload type integer
        in: query
        name: payloadType
        type: integer
      - description: Filter by multiple payload types, comma-separated e.g. 2,4
        in: query
        name: payloadTypes
        type: string
      - description: Filter by payload type name (advert, grp_txt, txt_msg, trace,
          anon_req)
        in: query
        name: payloadTypeName
        type: string
      - description: Filter by route type (0=transport_flood, 1=flood, 2=direct, 3=transport_direct)
        in: query
        name: routeType
        type: integer
      - description: Filter by multiple route types, comma-separated e.g. 0,1
        in: query
        name: routeTypes
        type: string
      - description: Filter by single IATA code (case-insensitive)
        in: query
        name: iata
        type: string
      - description: Filter by multiple IATA codes, comma-separated e.g. YVR,YYJ
        in: query
        name: iatas
        type: string
      - description: 'Filter by transport scope name e.g. %23bc (URL-encoded #bc)'
        in: query
        name: scope
        type: string
      - description: Filter by multiple transport scope names, comma-separated e.g.
          %23bc,%23west
        in: query
        name: scopes
        type: string
      - description: Filter by region ID, expands to member IATAs
        in: query
        name: regionId
        type: integer
      - description: Filter by region slug, expands to member IATAs
        in: query
        name: region
        type: string
      - description: Filter by first_heard_at >= since (epoch ms)
        in: query
        name: since
        type: integer
      - description: Filter by first_heard_at <= until (epoch ms)
        in: query
        name: until
        type: integer
      - description: epoch ms of last item for pagination; last_heard_at, or site-local
          heard_at when iatas is set
        in: query
        name: cursor
        type: integer
      - description: Max results (default 20, max 100)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_MeshCore-Beacon_beacon-server_internal_api.Page-github_com_MeshCore-Beacon_beacon-server_internal_api_PublicPacketSummary'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_api_handlers.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_api_handlers.APIError'
      summary: List packets
      tags:
      - Packets
  /packets/{packetHash}:
    get:
      parameters:
      - description: Packet hash (hex)
        in: path
        name: packetHash
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_MeshCore-Beacon_beacon-server_internal_api.PublicPacket'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_api_handlers.APIError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_api_handlers.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_api_handlers.APIError'
      summary: Get full packet detail. For trace packets (payloadType=9), includes
        resolvedRoute
      tags:
      - Packets
  /packets/backfill:
    get:
      parameters:
      - description: Return packets with observations after this ID (use last WS event
          observation ID)
        in: query
        name: afterObservationId
        required: true
        type: integer
      - description: Filter by payload type integer
        in: query
        name: payloadType
        type: integer
      - description: Filter by payload type name
        in: query
        name: payloadTypeName
        type: string
      - description: Filter by route type
        in: query
        name: routeType
        type: integer
      - description: Filter by IATA code(s), comma-separated
        in: query
        name: iatas
        type: string
      - description: Filter by region slug
        in: query
        name: region
        type: string
      - description: Filter by region ID
        in: query
        name: regionId
        type: integer
      - description: Filter by transport scope name
        in: query
        name: scope
        type: string
      - description: Max results (default 20, max 100)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_MeshCore-Beacon_beacon-server_internal_api.PublicPacketSummary'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_api_handlers.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_api_handlers.APIError'
      summary: Backfill packets after a given observation ID
      tags:
      - Packets
...
  /routes:
    get:
      parameters:
      - description: Filter by IATA code
        in: query
        name: iata
        type: string
      - description: Filter by exact hop count
        in: query
        name: hopCount
        type: integer
      - description: Epoch ms timestamp of last item for pagination
        in: query
        name: cursor
        type: integer
      - description: Max results (default 20, max 100)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_MeshCore-Beacon_beacon-server_internal_api.KnownRoute'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_api_handlers.APIError'
      summary: List known routes
      tags:
      - Routes
  /routes/cross:
    get:
      parameters:
      - description: Source node hash prefix (hex)
        in: query
        name: fromHash
        required: true
        type: string
      - description: Source IATA code
        in: query
        name: fromIata
        required: true
        type: string
      - description: Destination node hash prefix (hex)
        in: query
        name: toHash
        required: true
        type: string
      - description: Destination IATA code
        in: query
        name: toIata
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_MeshCore-Beacon_beacon-server_internal_api.CrossIATARoute'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_api_handlers.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_api_handlers.APIError'
      summary: Search for routes that cross IATA boundaries
      tags:
      - Routes
  /routes/search:
    get:
      parameters:
      - description: IATA code to search within
        in: query
        name: iata
        required: true
        type: string
      - description: Source node hash prefix (hex)
        in: query
        name: from
        required: true
        type: string
      - description: Destination node hash prefix (hex)
        in: query
        name: to
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_MeshCore-Beacon_beacon-server_internal_api.KnownRoute'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_api_handlers.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_api_handlers.APIError'
      summary: Search known routes by source and destination hash
      tags:
      - Routes
...
schemes:
- http
- https
swagger: "2.0"
...